+++ e2e-tests/development-runtime/gatsby-node.js
@@ -15,7 +15,7 @@ exports.onCreateNode = function onCreateNode({
 
       createNodeField({
         name: `slug`,
-        value: slug.replace("/", ""),
+        value: slug,
         node,
       })
       break
@@ -59,7 +59,7 @@ exports.createPages = async function createPages({
   data.posts.edges.forEach(({ node }) => {
     const { slug } = node.fields
     createPage({
-      path: `/${slug}`,
+      path: slug,
       component: blogPostTemplate,
       context: {
         slug: slug,
@@ -104,10 +104,34 @@ exports.createPages = async function createPages({
 }
 
 exports.onCreatePage = async ({ page, actions }) => {
-  const { createPage } = actions
+  const { createPage, createRedirect, deletePage } = actions
 
   if (page.path.match(/^\/client-only-paths/)) {
     page.matchPath = `/client-only-paths/*`
     createPage(page)
+  }
+
+  if (page.path === `/redirect-me/`) {
+    const toPath = `/pt${page.path}`
+
+    deletePage(page)
+
+    createRedirect({
+      fromPath: page.path,
+      toPath,
+      isPermanent: false,
+      redirectInBrowser: true,
+      Language: `pt`,
+      statusCode: 301,
+    })
+
+    createPage({
+      ...page,
+      path: toPath,
+      context: {
+        ...page.context,
+        lang: `pt`,
+      },
+    })
   }
 }