+++ packages/gatsby-telemetry/src/index.ts
@@ -38,6 +38,14 @@ export function trackCli(
   tags?: ITelemetryTagsPayload,
   opts?: ITelemetryOptsPayload
 ): void {
+  instance.trackCli(input, tags, opts)
+}
+
+export function captureEvent(
+  input: string | Array<string>,
+  tags?: ITelemetryTagsPayload,
+  opts?: ITelemetryOptsPayload
+): void {
   instance.captureEvent(input, tags, opts)
 }
 
@@ -84,9 +92,11 @@ export function addSiteMeasurement(event: string, obj): void {
 }
 
 export function expressMiddleware(source: string) {
-  return function (_req: Request, _res: Response, next): void {
+  return function (req: Request, _res: Response, next): void {
     try {
-      instance.trackActivity(`${source}_ACTIVE`)
+      instance.trackActivity(`${source}_ACTIVE`, {
+        userAgent: req.headers[`user-agent`],
+      })
     } catch (e) {
       // ignore
     }
@@ -103,6 +113,14 @@ export function setGatsbyCliVersion(version: string): void {
   instance.gatsbyCliVersion = version
 }
 
+export {
+  AnalyticsTracker,
+  IAggregateStats,
+  ITelemetryTagsPayload,
+  ITelemetryOptsPayload,
+  IDefaultTelemetryTagsPayload,
+}
+
 module.exports = {
   trackFeatureIsUsed,
   trackCli,
@@ -116,4 +134,6 @@ module.exports = {
   aggregateStats,
   addSiteMeasurement,
   expressMiddleware,
+  setDefaultComponentId,
+  setGatsbyCliVersion,
 }