+++ packages/gatsby/src/utils/webpack-utils.ts
@@ -8,7 +8,6 @@ import TerserPlugin from "terser-webpack-plugin"
 import MiniCssExtractPlugin from "mini-css-extract-plugin"
 import OptimizeCssAssetsPlugin from "optimize-css-assets-webpack-plugin"
 import ReactRefreshWebpackPlugin from "@pmmmwh/react-refresh-webpack-plugin"
-import isWsl from "is-wsl"
 import { getBrowsersList } from "./browserslist"
 
 import { GatsbyWebpackStatsExtractor } from "./gatsby-webpack-stats-extractor"
@@ -277,7 +276,6 @@ export const createWebpackUtils = (
         options: {
           stage,
           reactRuntime: jsxRuntimeExists ? `automatic` : `classic`,
-          // TODO add proper cache keys
           cacheDirectory: path.join(
             program.directory,
             `.cache`,
@@ -285,6 +283,7 @@ export const createWebpackUtils = (
             `babel`
           ),
           ...options,
+          rootDir: program.directory,
         },
         loader: require.resolve(`./babel-loader`),
       }
@@ -293,7 +292,6 @@ export const createWebpackUtils = (
     dependencies: options => {
       return {
         options: {
-          // TODO add proper cache keys
           cacheDirectory: path.join(
             program.directory,
             `.cache`,
@@ -401,9 +399,11 @@ export const createWebpackUtils = (
         // debugger to show the original code. Instead, the code
         // being evaluated would be much more helpful.
         sourceMaps: false,
-        cacheIdentifier: `${stage}---gatsby-dependencies@${
-          require(`babel-preset-gatsby/package.json`).version
-        }`,
+
+        cacheIdentifier: JSON.stringify({
+          browerslist: supportedBrowsers,
+          gatsbyPreset: require(`babel-preset-gatsby/package.json`).version,
+        }),
       }
 
       // TODO REMOVE IN V3
@@ -594,9 +594,6 @@ export const createWebpackUtils = (
     new TerserPlugin({
       // TODO add proper cache keys
       cache: path.join(program.directory, `.cache`, `webpack`, `terser`),
-      // We can't use parallel in WSL because of https://github.com/gatsbyjs/gatsby/issues/6540
-      // This issue was fixed in https://github.com/gatsbyjs/gatsby/pull/12636
-      parallel: !isWsl,
       exclude: /\.min\.js/,
       sourceMap: true,
       terserOptions: {