+++ packages/gatsby-plugin-styled-components/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-styled-components@3.3.13...gatsby-plugin-styled-components@3.3.14) (2020-10-09)
+
+**Note:** Version bump only for package gatsby-plugin-styled-components
+
+## [3.3.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-styled-components@3.3.12...gatsby-plugin-styled-components@3.3.13) (2020-10-06)
+
+**Note:** Version bump only for package gatsby-plugin-styled-components
+
+## [3.3.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-styled-components@3.3.11...gatsby-plugin-styled-components@3.3.12) (2020-09-28)
+
+**Note:** Version bump only for package gatsby-plugin-styled-components
+
+## [3.3.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-styled-components@3.3.10...gatsby-plugin-styled-components@3.3.11) (2020-09-15)
+
+**Note:** Version bump only for package gatsby-plugin-styled-components
+
 ## [3.3.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-styled-components@3.3.9...gatsby-plugin-styled-components@3.3.10) (2020-07-09)
 
 **Note:** Version bump only for package gatsby-plugin-styled-components